pyproject.toml
[tool.poetry]
name = "svp_robot"
version = "1.0.0"
description = "This project is for Automating SVP testing Using Robot"
authors = []
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "~3.10"
selenium = "^4.7.2"
robotframework = "^6.0.2"
robotframework-seleniumlibrary = "^6.0.0"
robotframework-pabot = "^2.11.0"
robotframework-jsonlibrary = "^0.5"
robotframework-browser = "16.0.0"
robotframework-robocop = "^2.7.0"
allure-robotframework = "^2.12.0"
PdfReader = "0.1.12"
requests = "2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
